Address comments from Bouwe in PR #195
