+++ esmvaltool/preprocessor/_reformat.py
@@ -1,34 +1,7 @@
 """Simple interface to reformat and CMORize functions."""
 from ..interface_scripts.cmor_check import CMORCheck
 from ..interface_scripts.fixes.fix import Fix
-from ..interface_scripts.variable_info import CMIP5Info, CMIP6Info
-from ..interface_scripts.data_finder import _CFG
-
-
-def _read_cmor_tables():
-    tables = {}
-
-    for table in _CFG.keys():
-        project = _CFG[table]
-
-        if 'cmor_table' in project:
-            table_path = project['cmor_table']
-        else:
-            table_path = None
-
-        if 'cmor_type' in project:
-            cmor_type = project['cmor_type']
-        else:
-            cmor_type = table
-
-        if cmor_type == 'CMIP5':
-            tables[table] = CMIP5Info(table_path)
-        elif cmor_type == 'CMIP6':
-            tables[table] = CMIP6Info(table_path)
-    return tables
-
-
-CMOR_TABLES = _read_cmor_tables()
+from ..interface_scripts.variable_info import CMOR_TABLES
 
 
 def _get_cmor_checker(table,
